Implementerade Log metoder åt customer för att kunna se tidigare deposits och withdrawls
